+++ CHANGELOG.md
@@ -11,6 +11,18 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 * Add debug info for resource usage
 
 
+### Fixed
+
+* `nh.api.Optimizer.coerce` now correctly respects lr and other params for
+  torch-optimizer optimizers.
+
+* MultiLayerPerceptronNd has several changes / fixes
+    - (BREAKING) always set `bias=False` in the hidden layers, this is now fixed. 
+    - dropout no longer is added if set to None
+    - option to apply a nonlineary / normalization to the output layer
+    - `output_shape_for` and `receptive_field_for` are now implemented via `_analytic_forward`
+
+
 ## Version 0.5.12 - Released 2020-12-07
 
 ### Fixed